- merge duplicate activities: login and sign up, and delete the reserve.
