+++ app/src/main/AndroidManifest.xml
@@ -14,22 +14,20 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SignUpActivity"></activity>
-        <activity android:name=".LoginActivity" />
         <activity android:name=".HomePage" />
         <activity android:name=".FriendShazamzamsActivity"></activity>
         <activity android:name=".FriendProfileActivity" />
         <activity android:name=".FriendsPageActivity" />
         <activity android:name=".MyProfileActivity" ></activity>
-        <activity android:name=".SignUp" />
+        <activity android:name=".SignUpActivity"/>
         <activity android:name=".AccountActivity" />
-        <activity android:name=".AuthActivity" />
-        <activity android:name=".HomePageActivity">
+        <activity android:name=".LoginActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".HomePageActivity"></activity>
     </application>
 
 </manifest>
