--- app/src/main/java/com/example/secondproject/PianoActivity.java
@@ -1,14 +1,0 @@
-package com.example.secondproject;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class PianoActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_piano);
-    }
-}
