+++ app/src/main/res/layout/activity_login.xml
@@ -37,12 +37,49 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="@+id/signUp_imageView2" />
 
+    <ImageButton
+        android:id="@+id/loginBtn"
+        android:layout_width="110dp"
+        android:layout_height="55dp"
+        android:layout_marginBottom="250dp"
+        android:background="@android:color/transparent"
+        android:onClick="launchHomePageActivity"
+        android:scaleType="centerCrop"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/password_editText"
+        app:srcCompat="@drawable/lets_go_button" />
+
+    <ImageButton
+        android:id="@+id/signUp_imageButton"
+        android:layout_width="110dp"
+        android:layout_height="55dp"
+        android:layout_marginTop="17dp"
+        android:background="@android:color/transparent"
+        android:onClick="launchSignUpActivity"
+        android:scaleType="centerCrop"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/googleBtn"
+        app:srcCompat="@drawable/sign_up_image" />
+
+    <ImageView
+        android:id="@+id/decoration_background"
+        android:layout_width="wrap_content"
+        android:layout_height="250dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:srcCompat="@drawable/decorations_button" />
+
+
     <EditText
         android:id="@+id/email_editText"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_margin="20dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="@dimen/global_top_margin"
         android:ems="10"
         android:hint="Email..."
         android:inputType="textEmailAddress"
@@ -62,41 +99,14 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/email_editText" />
 
-    <ImageButton
-        android:id="@+id/letsGo_imageButton"
-        android:layout_width="110dp"
-        android:layout_height="55dp"
-        android:layout_marginBottom="200dp"
-        android:background="@android:color/transparent"
-        android:scaleType="centerCrop"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/password_editText"
-        app:srcCompat="@drawable/lets_go_button"
-        android:onClick="launchHomePageActivity"/>
-
-    <ImageButton
-        android:id="@+id/signUp_imageButton"
-        android:layout_width="110dp"
-        android:layout_height="55dp"
-        android:layout_marginTop="@dimen/global_top_margin"
-        android:background="@android:color/transparent"
-        android:scaleType="centerCrop"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/letsGo_imageButton"
-        app:srcCompat="@drawable/sign_up_image"
-        android:onClick="launchSignUpActivity" />
-
-    <ImageView
-        android:id="@+id/decoration_background"
+    <com.google.android.gms.common.SignInButton
+        android:id="@+id/googleBtn"
         android:layout_width="wrap_content"
-        android:layout_height="250dp"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/global_top_margin"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/decorations_button" />
-
+        app:layout_constraintTop_toBottomOf="@+id/loginBtn"></com.google.android.gms.common.SignInButton>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
+
